#25: change to typescript, added background and hud
